bot/pyproject.toml
# bot/pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = "bot of tgminiapptemplate"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.9.0",
    "httpx>=0.27.0",
    "aiogram>=3.12.0",
    "python-decouple>=1.0.1",
    "pydantic>=2.9.0",
    "rich>=13.9.2",
    "python-dotenv"
]

[dependency-groups]
dev = [
    "watchdog>=6.0.0",
]